+++ src/BooleanController.js
@@ -14,6 +14,7 @@ export default class BooleanController extends Controller {
 
 		this.$input.addEventListener( 'change', () => {
 			this.setValue( this.$input.checked );
+			this._callOnFinishChange();
 		} );
 
 		this.$disable = this.$input;