+++ Changelog.md
@@ -1,3 +1,13 @@
+# 0.13.0
+
+28.56kb, 8.24kb gzipped
+
+- Added support for `onFinishChange`.
+- The error message when `gui.add` fails is easier to read.
+- `gui.load` ignores empty objects instead of failing.
+- Fixed a bug that caused mouse wheels to change values too slowly.
+- Reverted `listen()` fix from 0.11.0. ([Note](https://github.com/georgealways/lil-gui/pull/11/files#r748852628))
+
 # 0.12.0
 
 28.11kb, 8.24kb gzipped